refactor: Turn MultiValidator into AjaxCompositeValidator. TUM001-114
This is a subclass of the new CompositeValidator class and replaces that class in the default getCMSCompositeValidator method. Ommitting this refactoring would cause an issue where the validate method of the MultiValidator wouldn't have been called by CompositeValidator (see https://github.com/silverstripe/silverstripe-framework/issues/9868 for more information).

Effectively this validator now only adds AJAX validation to the CompositeValidator implementation.