refactor: Change inheritance of validators for configuration purposes.
With ChecksIfFieldHasValue as a trait, it is hard to take advantage of the updateFieldHasValue extension point without applying the Extension class that implements it to all classes that use ChecksIfFieldHasValue individually.

By turning ChecksIfFieldHasValue into an abstract class, we allow configuration and extensions to be inherited. MultiFieldValidator then needs to become a trait so that classes can take advantage of both it and the new FieldHasValueValidator class.