feat: Add tests for GridFieldMessagesExtension.
These tests should function correctly even after Silverstripe Framework
merges in the pull request fixing the original problem, since that adds
the messages to the raw output, not to the field description.
Therefore this test will be valid for older versions of Silverstripe
even after it's officially fixed in later versions.